+++ Framework/Source/Sample.cpp
@@ -42,6 +42,8 @@
 
 namespace Falcor
 {
+    static std::string kMonospaceFont = "monospace";
+
     void Sample::handleWindowSizeChange()
     {
         if (!gpDevice) return;
@@ -384,12 +386,14 @@ namespace Falcor
 
             if (gProfileEnabled)
             {
+                mpGui->setActiveFont(kMonospaceFont);
                 mpGui->pushWindow("Profiler", 650, 200, 10, 300);
                 // Stop the timer
                 Profiler::endEvent("renderGUI");
                 mpGui->addText(Profiler::getEventsString().c_str());
                 Profiler::startEvent("renderGUI");
                 mpGui->popWindow();
+                mpGui->setActiveFont("");
             }
 
             mpGui->render(mpRenderContext.get(), mFrameRate.getLastFrameTime());
@@ -528,7 +532,11 @@ namespace Falcor
 
     void Sample::initUI()
     {
-        mpGui = Gui::create(mpBackBufferFBO->getWidth(), mpBackBufferFBO->getHeight());
+        float scaling = getDisplayScaleFactor();
+        mpGui = Gui::create(uint32_t(mpBackBufferFBO->getWidth()), uint32_t(mpBackBufferFBO->getHeight()), scaling);
+        mpGui->addFont(kMonospaceFont, "Framework/Fonts/consolab.ttf");
+        mSampleGuiHeight = (uint32_t)(mSampleGuiHeight * scaling);
+        mSampleGuiWidth = (uint32_t)(mSampleGuiWidth * scaling);
         mpTextRenderer = TextRenderer::create();
     }