+++ Framework/Source/Utils/Platform/OS.cpp
@@ -176,17 +176,6 @@ namespace Falcor
             return str;
         }
     }
-
-    bool readFileToString(const std::string& fullpath, std::string& str)
-    {
-        std::ifstream t(fullpath.c_str());
-        if ((t.rdstate() & std::ifstream::failbit) == 0)
-        {
-            str = std::string((std::istreambuf_iterator<char>(t)), std::istreambuf_iterator<char>());
-            return true;
-        }
-        return false;
-    }
     
     std::string getDirectoryFromFile(const std::string& filename)
     {